Done minor refactoring. 1) Renamed TripUnitOfWork to TripRepository 2) Renamed TripDatabaseSettings to DatabaseSettings
