+++ Tavisca.Tripster.Contracts/Service/ITripService.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Tavisca.Tripster.Core.Validation;
 using Tavisca.Tripster.Data.Models;
