+++ Tavisca.Tripster.Core/Validation/Validator.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Tavisca.Tripster.Core.Validation
 {
