+++ Tavisca.Tripster.Data/Models/Location.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Tavisca.Tripster.Data.Models
+﻿namespace Tavisca.Tripster.Data.Models
 {
     public class Location
     {
