+++ Tavisca.Tripster.Data/Models/Place.cs
@@ -1,9 +1,4 @@
-﻿using MongoDB.Bson.Serialization.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Tavisca.Tripster.Data.Models
+﻿namespace Tavisca.Tripster.Data.Models
 {
     public class Place
     {
