+++ Tavisca.Tripster.Data/Models/Stop.cs
@@ -1,7 +1,4 @@
-﻿using MongoDB.Bson.Serialization.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace Tavisca.Tripster.Data.Models
 {
