+++ Tavisca.Tripster.MongoDB/Repository/MongoRepository.cs
@@ -1,7 +1,6 @@
 ﻿using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Tavisca.Tripster.Contracts.Repository;
